FaceDetector
This application is the one of my 1st application but this application
are not finish because I have a lot of project.
this app can detect faces when you tap take a picture if you have a lot
of cam app this app question you what cam app that you want to use and
and you can used it once or always and after take a picture tap ok then
tap detect face, this app detect between eyes, actually nose line and
"kilay".